pyproject.toml
[project]
name = "moonlight-pixelrun-tools"
version = "1.0.0"
description = "Sprite generation tools for Moonlight: Pixel Run"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
generate-sprites = "scripts.generate_sprites:main"
